refactor : 코드 리뷰 사항 반영
