+++ app/(anon)/main/_components/guideResults/guideResultSummary/GuideResultSummary.tsx
