+++ app/(anon)/main/_components/guideResults/guideResultView/GuideResultView.styles.ts
