+++ app/(anon)/main/_components/guideResults/resultAccordion/ResultAccordion.styles.ts
