+++ app/(anon)/main/_components/guideResults/resultAccordion/ResultAccordion.tsx
